% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_ist}
\alias{add_ist}
\title{Add Series}
\usage{
add_ist(p, values, name, ...)
}
\arguments{
\item{p}{a chartist object}

\item{values}{values variable}

\item{...}{Additional parameter to pass to line}
}
\description{
Add series(s) to chart
}

