% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{xaxis_ist}
\alias{xaxis_ist}
\alias{yaxis_ist}
\title{Tweak axis for chartist visualisation}
\usage{
xaxis_ist(p, ..., type = "auto", suffix, prefix, ticks, onlyInteger, offset,
  scaleMinSpace, labelInterpolationFnc)

yaxis_ist(p, ..., type = "auto", suffix, prefix, ticks, onlyInteger, offset,
  scaleMinSpace, labelInterpolationFnc)
}
\description{
Tweak axis for chartist visualisation
}

