% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{axistitle_ist}
\alias{axistitle_ist}
\title{Add x axis title}
\usage{
axistitle_ist(p, title = "", textAnchor = "middle", flipTitle = FALSE,
  offsetx = 0, offsety = 0)
}
\arguments{
\item{p}{a chartist object.}

\item{title}{axis title.}

\item{textAnchor}{title position, defaults to \code{middle}.}

\item{flipTitle}{set to \code{TRUE} to flip title on x axis.}
}
\description{
Add x axis title
}

