% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{baropt_ist}
\alias{baropt_ist}
\title{Bar options}
\usage{
baropt_ist(p, ..., fullWidth = TRUE, centerBars = FALSE,
  stackBars = FALSE, distributeSeries = FALSE, seriesBarDistance, width,
  height, low, high, chartPadding)
}
\arguments{
\item{p}{a chartist object.}

\item{...}{additional options.}

\item{fullWidth}{When set to true, the last grid line on the x-axis is not
drawn and the chart elements will expand to the full available width of the
chart. For the last label to be drawn correctly you might need to add chart
padding or offset the last label with a draw event handler.}

\item{centerBars}{set to \code{TRUE} to draw bars on the grid lines,
\code{FALSE} to draw on grid lines.}

\item{stackBars}{whether to stack bars, default to \code{FALSE}.}

\item{distributeSeries}{distributed serie along x axis only works for 1
serie, defaults to \code{FALSE}.}

\item{seriesBarDistance}{distance in pixels between bar groups.}

\item{width}{set fixed width in pixels or percent (i.e.: \code{300px}, \code{70\%}).}

\item{height}{set fixed height in pixels or percent
(i.e.: \code{300px}, \code{70\%}).}

\item{low}{lower bound of value.}

\item{high}{higher bound of value.}

\item{chartPadding}{padding around chart div.}
}
\description{
Pass options to \code{bar} type.
}
\examples{
mtcars$models <- rownames(mtcars)

mtcars[1:10,] \%>\%
    chart_ist(x = "models", type = "bar") \%>\%
    add_ist("cyl") \%>\%
    add_ist("carb") \%>\%
    add_ist("gear") \%>\%
    baropt_ist(stackBars = TRUE)

mtcars[1:10,] \%>\%
    chart_ist(x = "models", type = "bar") \%>\%
    add_ist("cyl") \%>\%
    add_ist("carb") \%>\%
    add_ist("gear") \%>\%
    baropt_ist(seriesBarDistance = 10, centerBars = TRUE)

}

