% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chartist.R
\name{chartist}
\alias{chartist}
\alias{chartist_}
\title{Draw a Chartist.js chart}
\usage{
chartist_(data, x_labels, select_cols)

chartist(data, x_label_colname, ...)
}
\arguments{
\item{data}{data.frame containing x-axis label and data series}

\item{x_labels}{x-axis labels, or the column name which contains x-axis labels}

\item{select_cols}{column name(s) of data}

\item{...}{column name(s) of data. If not specified, use all column except for \code{label_colname}}
}
\description{
Draw a Chartist.js chart. \code{chartist} is NSE version and \code{chartist_} is SE version.
You can specify column the same way as \code{dplyr}'s \code{select} does (The trick is very simple;
actually \code{select} is used inside).

NSE Version of function to create chartist plot
}
\examples{
data <- data.frame(
  day = paste0("day", 1:10),
  A   = runif(10, 0, 10),
  B   = runif(10, 0, 10),
  C   = runif(10, 0, 10)
)

# NSE version
chartist(data, day)

# These are the same
chartist(data, day, A, B)
chartist(data, day, A:B)
chartist(data, day, -C)

# SE version
chartist_(data, "day", c("A", "B"))

# SE version is usuful when you use variables
lab <- "day"
col <- c("A", "B")
chartist_(data, lab, col)

# With using SE version, labels can be specified directly
chartist_(data, rep(c("Odd", "Even"), 5), "A")
}
\seealso{
\url{http://gionkunz.github.io/}
}

