% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{danim_ist}
\alias{danim_ist}
\title{Animate donut chart}
\usage{
danim_ist(p)
}
\arguments{
\item{p}{a chartist object.}
}
\description{
Out of the box animation for donut chart.
}
\examples{
mtcars[1:5,] \%>\%
    chart_ist(x = hp, type = "pie") \%>\%
    add_ist(disp) \%>\%
    pieopt_ist(donut = TRUE, showLabel = TRUE) \%>\%
    danim_ist()

}

