% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_ist}
\alias{draw_ist}
\title{Add on draw function}
\usage{
draw_ist(p, fun)
}
\arguments{
\item{p}{a chartist object}

\item{fun}{Javascript function, must be wrapped in \code{htmlwidgets::JS()}}
}
\description{
Add on draw function
}

