% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{ganim_ist}
\alias{ganim_ist}
\title{Animate grid}
\usage{
ganim_ist(p)
}
\arguments{
\item{p}{a chartist object.}
}
\description{
Out of the box animation for the chart grid.
}
\examples{
mtcars \%>\%
    chart_ist(wt) \%>\%
    add_ist(disp) \%>\%
    add_ist(hp) \%>\%
    ganim_ist()

}

