% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gauge_ist}
\alias{gauge_ist}
\title{Make a gauge from a pie}
\usage{
gauge_ist(p, ...)
}
\arguments{
\item{p}{a chartist object}

\item{...}{additional parameters to pass to \code{\link{opt_ist}}}
}
\description{
Helper function to easily make a gauge chart from a pie
}
\examples{
mtcars[1:5,] \%>\%
    chart_ist(type = "pie") \%>\%
    add_ist("disp") \%>\%
    gauge_ist()

}

