% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{get_label_offset}
\alias{get_label_offset}
\title{possible patterns:
  1
  c(1, 2)
  list(x = 1, y = 2)}
\usage{
get_label_offset(labelOffset)
}
\description{
possible patterns:
  1
  c(1, 2)
  list(x = 1, y = 2)
}

