% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/javascript-functions.R
\name{JS_interp}
\alias{JS_interp}
\title{A Helper Function To Create JS Functions For labelInterporationFnc}
\usage{
JS_interp(interval = 1, offset = 0, prefix = NULL, suffix = NULL)
}
\arguments{
\item{interval}{interval between labels}

\item{offset}{offset to the first label}

\item{prefix}{prefix of labels (e.g. $)}

\item{suffix}{suffix of labels (e.g. kB)}
}
\description{
A Helper Function To Create JS Functions For labelInterporationFnc
}
\examples{
\dontrun{
#' set.seed(324)
data <- data.frame(
  day = 1:20,
  A   = runif(20, 0, 10),
  B   = runif(20, 0, 10),
  C   = runif(20, 0, 10),
  D   = runif(20, 0, 10)
  )

chartist(data, day) + Line(x_labelInterpolationFnc = JS_interp(4, offset = 2))

chartist(data, day) + Line(x_labelInterpolationFnc = JS_interp(4, prefix = "day "))
}
}

