% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{label_ist}
\alias{label_ist}
\title{Add point label}
\usage{
label_ist(p, position = "middle")
}
\arguments{
\item{p}{a chartist object.}

\item{position}{position of label}
}
\description{
Add data point labels to \code{line} type chart.
}

