% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{labelInterpolationFnc}
\alias{labelInterpolationFnc}
\title{bipolar}
\usage{
labelInterpolationFnc(x)
}
\arguments{
\item{x}{Label Interpolation Function to retrieve, see details for valid values}
}
\description{
bipolar
}
\details{
\itemize{
  \item{bipolar}
  \item{overlap_bars}
}
}

