% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{lanim_ist}
\alias{lanim_ist}
\title{Line and area chart animation}
\usage{
lanim_ist(p, type = c("line", "area"))
}
\arguments{
\item{p}{a chartist object.}

\item{type}{chart element type to animate (i.e.: line, or grid).}
}
\description{
out of the box animation for line and/or area chart
}

