% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{line_ist}
\alias{line_ist}
\title{Line options}
\usage{
line_ist(p, ..., fullWidth = TRUE, fillHoles = TRUE, showPoint = TRUE,
  showArea = FALSE, showLine = TRUE, lineSmoothing = "simple")
}
\description{
Line options
}

