% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{lineopt_ist}
\alias{lineopt_ist}
\title{Line options}
\usage{
lineopt_ist(p, ..., fullWidth = TRUE, fillHoles = TRUE, showPoint = TRUE,
  showArea = FALSE, showLine = TRUE, lineSmoothing = "simple")
}
\description{
Line options
}

