% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{loop_ist}
\alias{loop_ist}
\title{Loop animation}
\usage{
loop_ist(p, ms = 5000)
}
\arguments{
\item{p}{a chartist object.}

\item{ms}{Milliseconds for loop, defaults to \code{500}.}
}
\description{
Loop animation
}
\examples{
mtcars \%>\%
    chart_ist(qsec) \%>\%
    add_ist(hp) \%>\%
    lanim_ist() \%>\%
    loop_ist(6000)

}

