% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{name_ist}
\alias{name_ist}
\title{Get assigned name}
\usage{
name_ist(p)
}
\arguments{
\item{p}{a chartist object}
}
\description{
Get assigned name
}
\examples{
# default names
mtcars[1:20,] \%>\%
    chart_ist(x = qsec) \%>\%
    add_ist(disp) \%>\%
    add_ist(hp) \%>\%
    name_ist()

#assigned names
mtcars[1:20,] \%>\%
    chart_ist(x = qsec) \%>\%
    add_ist(disp, name = "SERIE1") \%>\%
    add_ist(hp, name = "hp") \%>\%
    name_ist()

}

