% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{opt_ist}
\alias{opt_ist}
\title{Configure options for chartist visualisation}
\usage{
opt_ist(p, ..., name, fillHoles = TRUE, showPoint = TRUE,
  showArea = FALSE, showLine = TRUE, lineSmoothing = "simple",
  fullWidth = TRUE, fillSmooth = TRUE, donut, donutWidth, startAngle, total,
  high, low, scaleMinSpace, referenceValue, stretch, seriesBarDistance,
  reverseData, horizontalBars)
}
\description{
Configure options for chartist visualisation
}

