% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{+.chartist}
\alias{plus}
\alias{update_chartist_options}
\title{Add options by + operand}
\usage{
\method{+}{chartist}(e1, e2)

update_chartist_options(chartist_plot, chartist_options)
}
\description{
Add options by + operand

Update options of Chartist plot
}

