% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartist.R
\name{renderChartist}
\alias{renderChartist}
\title{Widget render function for use in Shiny}
\usage{
renderChartist(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

