% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{resopts_ist}
\alias{resopts_ist}
\title{Configure responsive options for chartist visualisation}
\usage{
resopts_ist(p, ..., showPoint)
}
\description{
Configure responsive options for chartist visualisation
}

