% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{sanim_ist}
\alias{sanim_ist}
\title{Scatter animation}
\usage{
sanim_ist(p)
}
\arguments{
\item{p}{a chartist object.}
}
\description{
Out of the box animation for scatter plot.
}
\examples{
mtcars \%>\%
    chart_ist("wt") \%>\%
    add_ist("disp") \%>\%
    add_ist("hp") \%>\%
    scatter_ist() \%>\%
    sanim_ist()

}

