% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scatter_ist}
\alias{scatter_ist}
\title{Make a scatter plot from a line chart}
\usage{
scatter_ist(p, ...)
}
\arguments{
\item{p}{a chartist object}

\item{...}{additional parameters to pass to \code{\link{opt_ist}}}
}
\description{
Helper function to easily turn a line chart into a scatter plot
}
\examples{
mtcars[1:20,] \%>\%
    chart_ist(x = rownames(.)) \%>\%
    add_ist("disp") \%>\%
    scatter_ist()

}

