% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SVG_animate.R
\name{SVG_animate}
\alias{SVG_animate}
\title{SVG animation}
\usage{
SVG_animate(target = NULL, style = "opacity", easing = "easeOutQuart",
  delay = NULL, dur = 1000, interval = 5000, offset = NULL,
  relative = TRUE)
}
\description{
SVG animation
}

