% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{table_ist}
\alias{table_ist}
\title{Add table}
\usage{
table_ist(p, caption = "", seriesHeader = "", summary = "")
}
\arguments{
\item{p}{a chartist object.}

\item{caption}{table caption.}

\item{seriesHeader}{header for series.}

\item{summary}{table summary.}
}
\description{
Add a table below the chart with the accessibility plugin.
}
\keyword{internal}

