% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{thresh_ist}
\alias{thresh_ist}
\title{Add threshold}
\usage{
thresh_ist(p, threshold)
}
\arguments{
\item{p}{a chartist object.}

\item{threshold}{threshold, defaults to \code{mean} of first series passed with \code{\link{add_ist}}.}
}
\description{
Add threshold
}
\examples{
mtcars \%>\%
    chart_ist(disp) \%>\%
    add_ist(hp) \%>\%
    thresh_ist()

mtcars \%>\%
    chart_ist(disp) \%>\%
    add_ist(hp) \%>\%
    thresh_ist(220)

}

