% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{xaxis_ist}
\alias{xaxis_ist}
\title{Tweak X axis}
\usage{
xaxis_ist(p, ..., showLabel = TRUE, showGrid = TRUE, offset, labelOffset,
  scaleMinSpace)
}
\arguments{
\item{p}{a chartist object.}

\item{...}{additional options.}

\item{showLabel}{set to \code{FALSE} to hide labels.}

\item{showGrid}{set to \code{FALSE} to hide the grid.}

\item{offset}{offset axis.}

\item{labelOffset}{offset labels, takes \code{integer}, \code{vector} or
\code{list} (i.e.: \code{1}, \code{c(1,2)}, \code{list(x = 1, y = 2)}.}

\item{scaleMinSpace}{minimum height in pixel of scale.}
}
\description{
Tweak the x axis for chartist visualisation.
}

