% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{xtitle_ist}
\alias{xtitle_ist}
\title{Add x axis title}
\usage{
xtitle_ist(p, title = "", textAnchor = "middle", flipTitle = FALSE,
  offsetx = 0, offsety = 0)
}
\arguments{
\item{p}{a chartist object.}

\item{title}{axis title.}

\item{textAnchor}{title position, defaults to \code{middle}.}

\item{flipTitle}{set to \code{TRUE} to flip title on x axis.}

\item{offsetx}{offset x aligment of title.}

\item{offsety}{offsety y alignment of title.}
}
\description{
Add x axis title
}
\examples{
mtcars \%>\%
    chart_ist("disp") \%>\%
    add_ist(values = "hp") \%>\%
    xtitle_ist(title = "HP", offsety = 30)

}

