% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ylabel_ist}
\alias{ylabel_ist}
\title{Customize y axis labels}
\usage{
ylabel_ist(p, suffix = "", prefix = "", ...)
}
\arguments{
\item{p}{a chartist object.}

\item{suffix}{add suffix to label.}

\item{prefix}{add prefix to label.}

\item{...}{additional parameters to pass to y axis options.}
}
\description{
Customize y axis labels
}

