% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{ytitle_ist}
\alias{ytitle_ist}
\title{Add y axis title}
\usage{
ytitle_ist(p, title = "", flipTitle = FALSE, offsetx = 0, offsety = 0)
}
\arguments{
\item{p}{a chartist object.}

\item{title}{axis title.}

\item{flipTitle}{set to \code{TRUE} to flip title on x axis.}

\item{offsetx}{offset x aligment of title.}

\item{offsety}{offsety y alignment of title.}
}
\description{
Add y axis title
}
\examples{
mtcars \%>\%
    chart_ist(disp) \%>\%
    add_ist(values = hp) \%>\%
    ytitle_ist(title = "disp", offsetx = 30)

}

