% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{add_ist}
\alias{add_ist}
\title{Add Series}
\usage{
add_ist(p, x, y, ...)
}
\arguments{
\item{p}{a chartist object}

\item{x}{x variable}

\item{y}{y variable}

\item{...}{Additional parameter to pass to line}
}
\description{
Add series(s) to chart
}

