% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{anim_ist}
\alias{anim_ist}
\title{Animate chart}
\usage{
anim_ist(p, ..., type, target, begin, dur, from, to, easing)
}
\arguments{
\item{type}{chart element type to animate (i.e.: line, or grid).}

\item{target}{target of animation (i.e.: opacity).}

\item{begin}{begin index, \code{index}.}

\item{dur}{duration of animation in ms.}

\item{from}{start of animation.}

\item{to}{end of animation.}
}
\description{
Animate chart
}

