% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{anim_ist}
\alias{anim_ist}
\title{Animate chart}
\usage{
anim_ist(p, type, anim, begin, dur, from, to, easing, ...)
}
\arguments{
\item{p}{a chartist object.}

\item{type}{chart element type to animate (i.e.: line, or grid).}

\item{anim}{anim of animation (i.e.: opacity).}

\item{begin}{begin index, \code{index}.}

\item{dur}{duration of animation in ms.}

\item{from}{start of animation.}

\item{to}{end of animation.}

\item{easing}{apply easing.}

\item{...}{any other param to pass to \code{anim}.}
}
\description{
Animate chart
}

