% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{xaxis_ist}
\alias{xaxis_ist}
\alias{yaxis_ist}
\title{Tweak axis for chartist visualisation}
\usage{
xaxis_ist(p, ..., showLabel = TRUE, showGrid = TRUE, position, ticks,
  onlyInteger, offset, scaleMinSpace)

yaxis_ist(p, ..., showLabel = TRUE, showGrid = TRUE, position, ticks,
  onlyInteger, offset, scaleMinSpace)
}
\description{
Tweak axis for chartist visualisation
}

