% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartist.R
\name{chart_ist}
\alias{chart_ist}
\title{Initiate a chartist visualisation}
\usage{
chart_ist(data, type, width = NULL, height = NULL, elementId = NULL)
}
\description{
Transform data into a chartist visualisation
}

