% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartist.R
\name{chart_ist}
\alias{chart_ist}
\title{Initiate a chartist visualisation}
\usage{
chart_ist(data, x, type = "line", width = "100\%", height = "100\%",
  elementId = NULL)
}
\arguments{
\item{data}{\code{data.frame} containing data to plot.}

\item{x}{Must be passed for \code{line} and \code{bar} (\code{type}).}

\item{type}{type of charts, defaults to \code{line}, may also be \code{bar} or \code{pie}.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended, defaults to \code{100\%}.}

\item{elementId}{Use specific element id.}
}
\description{
Transform data into a chartist visualisation
}
\examples{
mtcars \%>\%
    chart_ist(cyl) \%>\%
    add_ist(hp)

}

