% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{draw_ist}
\alias{draw_ist}
\title{Pass function to on "draw"}
\usage{
draw_ist(p, fun)
}
\arguments{
\item{p}{a chartist object.}

\item{fun}{function, wrapped in \code{\link[htmlwidgets]{JS}}.}
}
\description{
Pass Javascript function to \code{chart.on("draw", your_fun)}.
}
\keyword{internal}

