% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{hover_ist}
\alias{hover_ist}
\title{Add hover tooltip}
\usage{
hover_ist(p)
}
\arguments{
\item{p}{a chartist object.}
}
\description{
Add hover tooltip
}
\details{
The tooltip displays series name which is passed in \code{\link{add_ist}}
}
\examples{
mtcars \%>\%
    chart_ist("disp") \%>\%
    add_ist(values = "hp", name = "Your Tooltip") \%>\%
    hover_ist()

mtcars \%>\%
    chart_ist("disp") \%>\%
    add_ist(values = "hp", name = "Your <span style='color:blue;'>Tooltip</span>") \%>\%
    hover_ist()

}

