% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Line.R
\name{Line}
\alias{Line}
\title{Line plot}
\usage{
Line(x_offset = NULL, x_labelOffset = NULL, x_showLabel = NULL,
  x_showGrid = NULL, x_labelInterpolationFnc = NULL, y_offset = NULL,
  y_labelOffset = NULL, y_showLabel = NULL, y_showGrid = NULL,
  y_labelInterpolationFnc = NULL, y_scaleMinSpace = NULL, width = NULL,
  height = NULL, showLine = NULL, showPoint = NULL, showArea = NULL,
  areaBase = NULL, lineSmooth = NULL, low = NULL, high = NULL,
  chartPadding = NULL, fullWidth = NULL, classNames = NULL)
}
\description{
The Chartist line chart can be used to draw Line or Scatter charts.
}

