% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{opt_ist}
\alias{opt_ist}
\title{Configure options for chartist visualisation}
\usage{
opt_ist(p, ..., name, width, height, low, high, chartPadding)
}
\arguments{
\item{p}{a chartist object.}

\item{...}{additional options.}

\item{name}{name the serie, see details (optional).}

\item{width}{set fixed width in pixels or percent (i.e.: \code{300px}, \code{70\%}).}

\item{height}{set fixed height in pixels or percent
(i.e.: \code{300px}, \code{70\%}).}

\item{low}{lower bound of value.}

\item{high}{higher bound of value.}

\item{chartPadding}{padding around chart div.}
}
\description{
Pass any option to any chart here. Also see options specific to your chart
type, \code{\link{pieopt_ist}}, \code{\link{baropt_ist}} or
\code{\link{lineopt_ist}}
}
\examples{
mtcars[1:10,] \%>\%
    chart_ist(x = rownames(.)) \%>\%
    add_ist("wt") \%>\%
    add_ist("drat") \%>\%
    opt_ist(chartPadding = 10, width = 700, height = 200)

mtcars[1:10,] \%>\%
    chart_ist(x = "qsec") \%>\%
    add_ist("disp", name = "DISP") \%>\%
    add_ist("hp", name = "HP") \%>\%
    opt_ist(name = "DISP", lineSmoothing = "step") \%>\%
    opt_ist(name = "HP",
        showArea = TRUE,
        showLine = FALSE,
        showPoint = FALSE)

}

