% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{opt_ist}
\alias{opt_ist}
\title{Configure options for chartist visualisation}
\usage{
opt_ist(p, ..., high, low, fullWidth = TRUE, scaleMinSpace, onlyInteger,
  referenceValue, stretch, showArea, showLine, axisX, axisY)
}
\description{
Configure options for chartist visualisation
}

