% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{peak_ist}
\alias{peak_ist}
\title{Turn bar chart into peak chart}
\usage{
peak_ist(p, size = 15)
}
\arguments{
\item{p}{a chartist object.}

\item{size}{point size.}
}
\description{
Turn bar chart into peak chart
}
\examples{
mtcars \%>\%
    chart_ist(qsec, type = "bar") \%>\%
    add_ist(hp) \%>\%
    peak_ist()

}

