% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_ist}
\alias{style_ist}
\title{Style your chart}
\usage{
style_ist(p, ...)
}
\arguments{
\item{p}{a chartist object.}

\item{...}{stuff}
}
\description{
Style your chart
}
\keyword{internal}

