% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_ist}
\alias{style_ist}
\title{Style your chart}
\usage{
style_ist(p, type, css = htmltools::css())
}
\arguments{
\item{p}{a chartist object.}

\item{type}{chart element type to animate (i.e.: line, or grid).}

\item{css}{css to wrap in \code{\link[htmltools]{css}}}
}
\description{
Style your chart
}
\examples{
# missing style defaults to chart type
mtcars \%>\%
    chart_ist(qsec, "line") \%>\%
    add_ist(mpg) \%>\%
    style_ist(css = htmltools::css(stroke = "rgba(0,0,0,0.4)", stroke.dasharray="20")) \%>\%
    style_ist("grid", htmltools::css(stroke = "black")) \%>\%
    style_ist(c("point"), htmltools::css(stroke = "green"))
    
mtcars \%>\%
    chart_ist(qsec, "line") \%>\%
    add_ist(hp, "hp") \%>\%
    add_ist(disp, "disp") \%>\%
    lineopt_ist("hp", showPoint = FALSE, showArea = TRUE, showLine = FALSE) \%>\%
    xaxis_ist(showLabel = FALSE) \%>\%
    style_ist(css = htmltools::css(stroke = "rgba(231,1,2,0.4)", stroke.dasharray="20")) \%>\%
    style_ist("area", htmltools::css(stroke = "black", fill = "rgba(0,,0,0.9)")) \%>\%
    style_ist(c("point"), htmltools::css(stroke = "red"))

}

