% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{thresh_ist}
\alias{thresh_ist}
\title{Add threshold}
\usage{
thresh_ist(p, threshold)
}
\arguments{
\item{p}{a chartist object.}

\item{threshold}{position of label, defaults to \code{mean} of first series.}
}
\description{
Add threshold
}

